Removed make target "clean-doc" to clean only Doxygen outputs.
"clean-doc" provides cleaning only Doxygen outputs
without cleaning executable binaries and more.
But it is more useful to clean binaries without cleaning doc
(because updating doc is not more frequent than binaries),
and it is not provied.
This is a little confusing,
so unify targets to cleaning outputs into "clean".